% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt}
\alias{fmt}
\title{Decorate Text Printed to Console Using ANSI Escape Codes}
\usage{
fmt(x, fg = NA, bg = NA, bf = FALSE, it = FALSE, ul = FALSE)
}
\arguments{
\item{x}{A character vector giving the text to format}

\item{fg}{What color should the foreground (text) be?
Anything R recognizes as a color can be specified.
System colors can be accessed by the numbers 0-15,
and similarly for the 256-color grayscale block and the numbers 232-255.
If \code{NA} (the default), the foreground color is not changed.}

\item{bg}{What color should the background be?
Anything R recognizes as a color can be specified.
System colors can be accessed by the numbers 0-15,
and similarly for the 256-color grayscale block and the numbers 232-255.
If \code{NA} (the default), the background color is not changed.}

\item{bf}{Should the text be boldface? The default is \code{FALSE}.}

\item{it}{Should the text be italic? The default is \code{FALSE}.}

\item{ul}{Should the text be underlined? The default is \code{FALSE}.}
}
\value{
The text encased in the relevant ANSI escape code.
}
\description{
ANSI escape codes can be used to change the foreground color, background
color, and other graphical features of text printed to the console
}
\examples{
cat(fmt("Normal text"))
cat(fmt("fg = 'system red' (which may not be red)", fg = 9))
cat(fmt("fg = #ffd700 (close to Tango palette Butter)", fg = "#edd400"))
cat(fmt("fg = navy, bg = sky blue", fg = "#000080", bg = "#87ceeb"))
cat(fmt("bold", bf = TRUE))
cat(fmt("bold and italic", bf = TRUE, it = TRUE))
cat(fmt("bold, italic, and underlined", bf = TRUE, it = TRUE, ul = TRUE))

}
