% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get256color.R
\name{get256color}
\alias{get256color}
\title{Find the closest color in the 256 color system}
\usage{
get256color(color)
}
\arguments{
\item{color}{Anything that base::col2rgb() recognizes as a color,
or the numbers 0 to 255 inclusive}
}
\value{
A named numeric vector of the translated colors.
The elements are the number from 0 to 255 identifying the color.
(If the number of a 256 color is given, it stays unchanged).
The names are the Xterm names for the colors.
}
\description{
The 256 color system is made of an initial 16 terminal-dependent colors,
a 6 by 6 by 6 block of combinations of red, green, and blue values,
and 24 grayscale colors. This function respects selection of system or
grayscale block colors but translates any other color into the closest
match in the set of 256 colors.
}
\examples{
get256color(c("black", "white", "#bf5700", rgb(0.1, 0.3, 0.9), 3, 244))

}
