% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palette256}
\alias{palette256}
\title{Xterm / 256 Colors}
\format{
A named numeric vector with values \code{0:255} whose names are the
corresponding Xterm color names.
}
\source{
https://www.ditig.com/256-colors-cheat-sheet
}
\usage{
palette256
}
\description{
A vector containing the number
}
\keyword{datasets}
